backend/requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
psycopg2==2.8.5
Pillow==7.2.0
